+++ src/auth.ts
@@ -55,6 +55,17 @@ export const { handlers, signIn, signOut, auth } = NextAuth(async () => {
     session: {
       strategy: "jwt",
     },
+    cookies: {
+      sessionToken: {
+        name: `authjs.session-token`,
+        options: {
+          httpOnly: true,
+          sameSite: "lax",
+          path: "/",
+          secure: process.env.NODE_ENV === "production",
+        },
+      },
+    },
     providers: [
       Nodemailer({
         server: {